Connected Application Analytics (#379)
* Update tests, builds and doc (#318)

* rebased with bwc tests


* updated bwc tests


* added release notes


* Rename trace-analytics to observability (#341)


* Event analytics unit tests (#342)

* tests


* updated some snapshots


* resolved few failing tests


* few modifications


* Update service map parsing results for testing (#345)


* bumping version to 1.2.2 (#346)

* bumping version to 1.2.2


* update PR in release notes


* updating readme and badges (#352)


* Application Analytics (#299)

* Add database schema for application


* Finished front end for Application overview

* Finished application detail page tabs

* WIP: Overview Page

* Rough sketch of App Analytics UI

* Create dummy page

* Create app complete. Stabilizing dashboard component.

* Update to 1.2 Observability


* notebooks internal error


* Address comments on PR: copyright headers, indentation, unnecessary render props


* Set max width of app and event


* Remove optional after description


* Change to singular


* Remove count badge for log source


* #290: Change form row label to ppl base query


* #291: Change description and help text for log source


* Pass down proper props


* Resolve gradle error and module not found error


* Resolve kotlin errors


* Fix parsers


* Add praseItemList


* Camelcase fields


* Remove whitespace, add copyright


* #292: Add autocomplete to Log Source accordion


* Lexicographic kotlin import


* Add newline at end of files


* #293: Add service map to create page


* #304: Activate Clear All button for services


* #305: Add button to clear base query


* opensearch-project#295: Add eui combo box for trace groups


* Separate out configuration renders


* debug adding filters traces


* #296: Add traces table to config


* Change from tsx to ts


* opensearch-project#309: Add page props and add app specific filters


* #308: Add button to clear trace groups


* #311: Allow services and traces to be selected


* Remove link to traces on table


* Disable clear all if nothing selected


* disable clear all when no log source


* Remove comment, add style to constant, temporarily remove availability


* Address PR comments


* Revert type assignment


* Update tests, builds and doc (#318)

* rebased with bwc tests


* updated bwc tests


* added release notes


* Fix errors and address comments


* #319: Disable create until required fields are filled out


* #329: Add missing field tool tip


* Remove unnecessary imports


* #320: Add clear modal for friction



* Fix merge conflict


* Update Workflow (#360)

* updated snapshot in workflow to 1.2.3


* replaced variable with 1.2.3-snapshot


* revert string to variable


* Application Analytics (#299)

* Add database schema for application


* Finished front end for Application overview

* Finished application detail page tabs

* WIP: Overview Page

* Rough sketch of App Analytics UI

* Create dummy page

* Create app complete. Stabilizing dashboard component.

* Update to 1.2 Observability


* notebooks internal error


* Address comments on PR: copyright headers, indentation, unnecessary render props


* Set max width of app and event


* Remove optional after description


* Change to singular


* Remove count badge for log source


* #290: Change form row label to ppl base query


* #291: Change description and help text for log source


* Pass down proper props


* Resolve gradle error and module not found error


* Resolve kotlin errors


* Fix parsers


* Add praseItemList


* Camelcase fields


* Remove whitespace, add copyright


* #292: Add autocomplete to Log Source accordion


* Lexicographic kotlin import


* Add newline at end of files


* #293: Add service map to create page


* #304: Activate Clear All button for services


* #305: Add button to clear base query


* opensearch-project#295: Add eui combo box for trace groups


* Separate out configuration renders


* debug adding filters traces


* #296: Add traces table to config


* Change from tsx to ts


* opensearch-project#309: Add page props and add app specific filters


* #308: Add button to clear trace groups


* #311: Allow services and traces to be selected


* Remove link to traces on table


* Disable clear all if nothing selected


* disable clear all when no log source


* Remove comment, add style to constant, temporarily remove availability


* Address PR comments


* Revert type assignment


* Update tests, builds and doc (#318)

* rebased with bwc tests


* updated bwc tests


* added release notes


* Fix errors and address comments


* #319: Disable create until required fields are filled out


* #329: Add missing field tool tip


* Remove unnecessary imports


* #320: Add clear modal for friction



* Application Analytics (#299)

* Add database schema for application


* Finished front end for Application overview

* Finished application detail page tabs

* WIP: Overview Page

* Rough sketch of App Analytics UI

* Create dummy page

* Create app complete. Stabilizing dashboard component.

* Update to 1.2 Observability


* notebooks internal error


* Address comments on PR: copyright headers, indentation, unnecessary render props


* Set max width of app and event


* Remove optional after description


* Change to singular


* Remove count badge for log source


* #290: Change form row label to ppl base query


* #291: Change description and help text for log source


* Pass down proper props


* Resolve gradle error and module not found error


* Resolve kotlin errors


* Fix parsers


* Add praseItemList


* Camelcase fields


* Remove whitespace, add copyright


* #292: Add autocomplete to Log Source accordion


* Lexicographic kotlin import


* Add newline at end of files


* #293: Add service map to create page


* #304: Activate Clear All button for services


* #305: Add button to clear base query


* opensearch-project#295: Add eui combo box for trace groups


* Separate out configuration renders


* debug adding filters traces


* #296: Add traces table to config


* Change from tsx to ts


* opensearch-project#309: Add page props and add app specific filters


* #308: Add button to clear trace groups


* #311: Allow services and traces to be selected


* Remove link to traces on table


* Disable clear all if nothing selected


* disable clear all when no log source


* Remove comment, add style to constant, temporarily remove availability


* Address PR comments


* Revert type assignment


* Update tests, builds and doc (#318)

* rebased with bwc tests


* updated bwc tests


* added release notes


* Fix errors and address comments


* #319: Disable create until required fields are filled out


* #329: Add missing field tool tip


* Remove unnecessary imports


* #320: Add clear modal for friction



* Fix merge conflict


* #321: Add clear modal for services (#355)


* #322: Add clear modal for traces (#356)


* Adds tests for application creation (#357)

* Remove unnecessary memoization


* Add tests for create page


* Remove conflict


* Add create route for app analytics


* #325: Return to app table on cancel


* Remove unnecessary import


* #340: Apps selectable, actions correctly disabled


* Fix mapping update


* #327: Rename application


* #376: Fetch all applications for home page


* Add exception catch


* More specific exception
:


* Log error instead of warn


* Remove generic exception


* Debug assert


* Fix type


* Fix Lint


* Address comments


* Move types to type file
