+++ dashboards-observability/common/types/app_analytics.ts
@@ -1,0 +1,23 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export interface optionType {
+  label: string;
+}
+
+export type ApplicationListType = {
+  name: string;
+  id: string;
+  dateCreated: string;
+  dateModified: string;
+};
+
+export type ApplicationType = {
+  name: string;
+  description: string;
+  query: string;
+  selectedServices: Array<string>;
+  selectedTraces: Array<string>;
+}
