+++ dashboards-observability/public/components/application_analytics/components/app_table.tsx
@@ -32,7 +32,7 @@ import { AppAnalyticsComponentDeps } from '../home';
 import { getCustomModal } from '../../custom_panels/helpers/modal_containers';
 import { getClearModal } from '../helpers/modal_containers';
 import { pageStyles } from '../../../../common/constants/shared';
-import { ApplicationListType } from '../../../../common/constants/application_analytics';
+import { ApplicationListType } from '../../../../common/types/app_analytics';
 
 interface AppTableProps extends AppAnalyticsComponentDeps {
     loading: boolean;
