+++ dashboards-observability/public/components/application_analytics/components/config_components/service_config.tsx
@@ -12,7 +12,7 @@ import DSLService from "public/services/requests/dsl";
 import React, { useState } from "react";
 import { useEffect } from "react";
 import { AppAnalyticsComponentDeps } from "../../home";
-import { optionType } from "../../../../../common/constants/application_analytics";
+import { optionType } from "../../../../../common/types/app_analytics";
 import { getClearModal } from "../../helpers/modal_containers";
 
 interface ServiceConfigProps extends AppAnalyticsComponentDeps {
