+++ dashboards-observability/public/components/application_analytics/components/config_components/trace_config.tsx
@@ -5,7 +5,7 @@
 
 import dateMath from '@elastic/datemath';
 import { EuiAccordion, EuiBadge, EuiButton, EuiComboBox, EuiFormRow, EuiOverlayMask, EuiSpacer, EuiText } from "@elastic/eui";
-import { optionType } from "../../../../../common/constants/application_analytics";
+import { optionType } from "../../../../../common/types/app_analytics";
 import { filtersToDsl } from "../../../trace_analytics/components/common/helper_functions";
 import { handleDashboardRequest } from "../../../trace_analytics/requests/dashboard_request_handler";
 import DSLService from "public/services/requests/dsl";
