+++ dashboards-observability/public/components/application_analytics/home.tsx
@@ -18,7 +18,8 @@ import TimestampUtils from 'public/services/timestamp/timestamp';
 import { handleIndicesExistRequest } from '../trace_analytics/requests/request_handler';
 import { ObservabilitySideBar } from '../common/side_nav';
 import { NotificationsStart } from '../../../../../src/core/public';
-import { optionType, APP_ANALYTICS_API_PREFIX, ApplicationListType, ApplicationType } from '../../../common/constants/application_analytics';
+import { APP_ANALYTICS_API_PREFIX } from '../../../common/constants/application_analytics';
+import { optionType, ApplicationListType } from '../../../common/types/app_analytics';
 import { isNameValid } from './helpers/utils';
 import { EuiGlobalToastList } from '@elastic/eui';
 import { Toast } from '@elastic/eui/src/components/toast/global_toast_list';
