+++ dashboards-observability/server/routes/application_analytics/app_analytics_router.ts
@@ -10,7 +10,8 @@ import {
   ResponseError,
   ILegacyScopedClusterClient,
 } from '../../../../../src/core/server';
-import { APP_ANALYTICS_API_PREFIX as API_PREFIX, ApplicationListType } from '../../../common/constants/application_analytics';
+import { APP_ANALYTICS_API_PREFIX as API_PREFIX } from '../../../common/constants/application_analytics';
+import { ApplicationListType } from '../../../common/types/app_analytics';
 import { AppAnalyticsAdaptor } from '../../../server/adaptors/application_analytics/app_analytics_adaptor';
 
 export function registerAppAnalyticsRouter(router: IRouter) {
@@ -173,39 +174,38 @@ export function registerAppAnalyticsRouter(router: IRouter) {
     }
   );
 
-// Delete applications
-router.delete(
-  {
-    path: `${API_PREFIX}/{appList}`,
-    validate: {
-      params: schema.object({
-        appList: schema.string(),
-      }),
+  // Delete applications
+  router.delete(
+    {
+      path: `${API_PREFIX}/{appList}`,
+      validate: {
+        params: schema.object({
+          appList: schema.string(),
+        }),
+      },
     },
-  },
-  async (
-    context,
-    request,
-    response
-  ): Promise<IOpenSearchDashboardsResponse<any | ResponseError>> => {
-    const opensearchClient: ILegacyScopedClusterClient = context.observability_plugin.observabilityClient.asScoped(
-      request
-    );
-    try {
-      const delResponse = await appAnalyticsBackend.deleteApp(
-        opensearchClient,
-        request.params.appList
+    async (
+      context,
+      request,
+      response
+    ): Promise<IOpenSearchDashboardsResponse<any | ResponseError>> => {
+      const opensearchClient: ILegacyScopedClusterClient = context.observability_plugin.observabilityClient.asScoped(
+        request
       );
-      return response.ok({
-        body: delResponse,
-      });
-    } catch (err: any) {
-      return response.custom({
-        statusCode: err.statusCode || 500,
-        body: err.message,
-      });
+      try {
+        const delResponse = await appAnalyticsBackend.deleteApp(
+          opensearchClient,
+          request.params.appList
+        );
+        return response.ok({
+          body: delResponse,
+        });
+      } catch (err: any) {
+        return response.custom({
+          statusCode: err.statusCode || 500,
+          body: err.message,
+        });
+      }
     }
-  }
-);
-
+  );
 }
